+++ packages/start-server-core/vite.config.ts
@@ -24,6 +24,12 @@ const fakeRouterEntry = fileURLToPath(
 const fakeStartManifest = fileURLToPath(
   new URL('./tests/fixtures/start-manifest.ts', import.meta.url),
 )
+const fakePluginAdapters = fileURLToPath(
+  new URL('./src/empty-plugin-adapters.ts', import.meta.url),
+)
+const fakeServerFnResolver = fileURLToPath(
+  new URL('./src/fake-start-server-fn-resolver.ts', import.meta.url),
+)
 
 export default mergeConfig(
   config,
@@ -36,6 +42,8 @@ export default mergeConfig(
         './src/createServerRpc.ts',
         './src/createSsrRpc.ts',
         './src/request-response.ts',
+        './src/constants.ts',
+        './src/virtual-modules.ts',
         './src/fake-start-server-fn-resolver.ts',
         './src/empty-plugin-adapters.ts',
       ],
@@ -53,6 +61,8 @@ export default mergeConfig(
         alias: {
           '#tanstack-start-entry': fakeStartEntry,
           '#tanstack-router-entry': fakeRouterEntry,
+          '#tanstack-start-plugin-adapters': fakePluginAdapters,
+          '#tanstack-start-server-fn-resolver': fakeServerFnResolver,
           'tanstack-start-manifest:v': fakeStartManifest,
         },
       },